+++ tests/src/server-side-encryption/main.php
@@ -23,6 +23,16 @@ final class <classname> extends main
 
 class main extends PHPUnit\Framework\TestCase
 {
+	protected static function array_to_env($array) {
+		$result = [];
+
+		foreach ($array as $key => $value) {
+			$result[] = "$key=$value";
+		}
+
+		return implode(" ", $result);
+	}
+
 	protected static function clear_dir($path) {
 		if (is_dir($path)) {
 			$content = scandir($path);
@@ -88,11 +98,12 @@ class main extends PHPUnit\Framework\TestCase
 
 	public function test() {
 		if (is_dir(__DIR__."/../../data/server-side-encryption/".static::VERSION."/test/")) {
-			define("TESTING",        true);
-			define("DATADIRECTORY",  __DIR__."/../../data/server-side-encryption/".static::VERSION."/test/");
-			define("INSTANCEID",     static::INSTANCEID);
-			define("SECRET",         static::SECRET);
-			define("USER_PASSWORDS", static::USER_PASSWORDS);
+			putenv("TESTING=".          "true");
+			putenv("DATADIRECTORY=".    __DIR__."/../../data/server-side-encryption/".static::VERSION."/test/");
+			putenv("INSTANCEID=".       static::INSTANCEID);
+			putenv("RECOVERY_PASSWORD=".static::RECOVERY_PASSWORD);
+			putenv("SECRET=".           static::SECRET);
+			putenv("USER_PASSWORDS=".   static::array_to_env(static::USER_PASSWORDS));
 
 			include(__DIR__."/../../../server-side-encryption/recover.php");
 
@@ -113,10 +124,12 @@ class main extends PHPUnit\Framework\TestCase
 
 	public function test_master() {
 		if (is_dir(__DIR__."/../../data/server-side-encryption/".static::VERSION."/master/")) {
-			define("TESTING",       true);
-			define("DATADIRECTORY", __DIR__."/../../data/server-side-encryption/".static::VERSION."/master/");
-			define("INSTANCEID",    static::INSTANCEID);
-			define("SECRET",        static::SECRET);
+			putenv("TESTING=".          "true");
+			putenv("DATADIRECTORY=".    __DIR__."/../../data/server-side-encryption/".static::VERSION."/master/");
+			putenv("INSTANCEID=".       static::INSTANCEID);
+			putenv("RECOVERY_PASSWORD=".static::RECOVERY_PASSWORD);
+			putenv("SECRET=".           static::SECRET);
+			putenv("USER_PASSWORDS=".   static::array_to_env(static::USER_PASSWORDS));
 
 			include(__DIR__."/../../../server-side-encryption/recover.php");
 
@@ -137,10 +150,12 @@ class main extends PHPUnit\Framework\TestCase
 
 	public function test_pubshare() {
 		if (is_dir(__DIR__."/../../data/server-side-encryption/".static::VERSION."/pubshare/")) {
-			define("TESTING",       true);
-			define("DATADIRECTORY", __DIR__."/../../data/server-side-encryption/".static::VERSION."/pubshare/");
-			define("INSTANCEID",    static::INSTANCEID);
-			define("SECRET",        static::SECRET);
+			putenv("TESTING=".          "true");
+			putenv("DATADIRECTORY=".    __DIR__."/../../data/server-side-encryption/".static::VERSION."/pubshare/");
+			putenv("INSTANCEID=".       static::INSTANCEID);
+			putenv("RECOVERY_PASSWORD=".static::RECOVERY_PASSWORD);
+			putenv("SECRET=".           static::SECRET);
+			putenv("USER_PASSWORDS=".   static::array_to_env(static::USER_PASSWORDS));
 
 			include(__DIR__."/../../../server-side-encryption/recover.php");
 
@@ -161,11 +176,12 @@ class main extends PHPUnit\Framework\TestCase
 
 	public function test_recovery() {
 		if (is_dir(__DIR__."/../../data/server-side-encryption/".static::VERSION."/recovery/")) {
-			define("TESTING",           true);
-			define("DATADIRECTORY",     __DIR__."/../../data/server-side-encryption/".static::VERSION."/recovery/");
-			define("INSTANCEID",        static::INSTANCEID);
-			define("RECOVERY_PASSWORD", static::RECOVERY_PASSWORD);
-			define("SECRET",            static::SECRET);
+			putenv("TESTING=".          "true");
+			putenv("DATADIRECTORY=".    __DIR__."/../../data/server-side-encryption/".static::VERSION."/recovery/");
+			putenv("INSTANCEID=".       static::INSTANCEID);
+			putenv("RECOVERY_PASSWORD=".static::RECOVERY_PASSWORD);
+			putenv("SECRET=".           static::SECRET);
+			putenv("USER_PASSWORDS=".   static::array_to_env(static::USER_PASSWORDS));
 
 			include(__DIR__."/../../../server-side-encryption/recover.php");
 
@@ -186,11 +202,12 @@ class main extends PHPUnit\Framework\TestCase
 
 	public function test_user() {
 		if (is_dir(__DIR__."/../../data/server-side-encryption/".static::VERSION."/user/")) {
-			define("TESTING",        true);
-			define("DATADIRECTORY",  __DIR__."/../../data/server-side-encryption/".static::VERSION."/user/");
-			define("INSTANCEID",     static::INSTANCEID);
-			define("SECRET",         static::SECRET);
-			define("USER_PASSWORDS", static::USER_PASSWORDS);
+			putenv("TESTING=".          "true");
+			putenv("DATADIRECTORY=".    __DIR__."/../../data/server-side-encryption/".static::VERSION."/user/");
+			putenv("INSTANCEID=".       static::INSTANCEID);
+			putenv("RECOVERY_PASSWORD=".static::RECOVERY_PASSWORD);
+			putenv("SECRET=".           static::SECRET);
+			putenv("USER_PASSWORDS=".   static::array_to_env(static::USER_PASSWORDS));
 
 			include(__DIR__."/../../../server-side-encryption/recover.php");